// src/app/shared/models/customer.ts
export class Customer {

    customerID: number; 
    customerName: string;
    shortName: string;
    isActive: boolean;
    isNewCollect: boolean;
    unitsType: number;
    dateFormat: string;
    timeFormat: string;
    timeZone: string;
    isDaylightSavings: boolean;

    constructor() {
    }
}